Rewrote screen update algorithm
Rewrote screen update algorithm to not rely on fps.cpp and to allow checking for what parts of the screen have changed to try to mimic a conventional vnc server. The results weren’t as good as expected, so I added flags to toggle between the new algorithm and just sending the whole screen every update. Code has also been trimmed of unnecessary parts.